requirements.txt
Django>=3.2<3.3
django-dotenv==1.4.2
djangorestframework==3.12.4
Scrapy==2.5.0
scrapy-djangoitem==1.1.1